Move character control to new ActorController and child classes
