+++ Assets/Prefabs/TestPlayer.prefab
@@ -24,7 +24,7 @@ Transform:
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 241793939610492566}
-  m_LocalRotation: {x: 0.3973272, y: -0.20973475, z: 0.09379622, w: 0.88845074}
+  m_LocalRotation: {x: 0.39732724, y: -0.20973478, z: 0.09379625, w: 0.88845074}
   m_LocalPosition: {x: 4, y: 10, z: -8}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children:
@@ -61,7 +61,7 @@ MonoBehaviour:
     ModeOverride: 0
     LensShift: {x: 0, y: 0}
     GateFit: 2
-    m_SensorSize: {x: 1.9238477, y: 1}
+    m_SensorSize: {x: 2.1213872, y: 1}
   m_Transitions:
     m_BlendHint: 0
     m_InheritPosition: 0
@@ -82,6 +82,7 @@ GameObject:
   - component: {fileID: 4132459119649282267}
   - component: {fileID: 2308247825992639317}
   - component: {fileID: 1416680847249297647}
+  - component: {fileID: 9016991782078022471}
   m_Layer: 0
   m_Name: TestPlayer
   m_TagString: Untagged
@@ -117,10 +118,6 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: bf17ce64048efe647ad0678aca47a457, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  Position:
-    m_InternalValue: {x: 0, y: 0, z: 0}
-  Rotation:
-    m_InternalValue: {x: 0, y: 0, z: 0, w: 0}
 --- !u!114 &2308247825992639317
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -168,6 +165,23 @@ MonoBehaviour:
   PrefabHashGenerator: Player
   AlwaysReplicateAsRoot: 0
   DontDestroyWithOwner: 0
+--- !u!114 &9016991782078022471
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 4744603141306243263}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 4d332d5802a92464ea7176cdd09780b2, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  Position:
+    m_InternalValue: {x: 0, y: 0, z: 0}
+  Rotation:
+    m_InternalValue: {x: 0, y: 0, z: 0, w: 0}
+  IsWalking: 0
 --- !u!143 &8881669713636083125
 CharacterController:
   m_ObjectHideFlags: 0
