+++ Assets/_Code/Player/ActorController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5f6d7fa87c62ffb4f8ec2efd16d8e97f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
