+++ Assets/_Code/Player/Characters.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 9868ac979fc4fb34081b0eace9d8fada
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
