+++ Assets/_Code/Player/Characters/Guard.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 5b4489f7d68508d43af12d8fe58b3e6b
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
