+++ Assets/_Code/Player/Characters/Guard/GuardController.cs
@@ -1,0 +1,48 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class Guard : ActorController
+{
+    public override void Ability1()
+    {
+        throw new System.NotImplementedException();
+    }
+
+    public override void Attack()
+    {
+        throw new System.NotImplementedException();
+    }
+
+    public override void Emote1()
+    {
+        throw new System.NotImplementedException();
+    }
+
+    public override void Emote2()
+    {
+        throw new System.NotImplementedException();
+    }
+
+    public override void Emote3()
+    {
+        throw new System.NotImplementedException();
+    }
+
+    public override void Emote4()
+    {
+        throw new System.NotImplementedException();
+    }
+
+    // Start is called before the first frame update
+    void Start()
+    {
+        
+    }
+
+    // Update is called once per frame
+    override protected void Update()
+    {
+        base.Update();
+    }
+}
