+++ Assets/_Code/Player/Characters/Guard/GuardController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4d332d5802a92464ea7176cdd09780b2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
